+++ lib/active_currency/cacheable_store.rb
@@ -4,6 +4,8 @@ module ActiveCurrency
   # Mixin to add caching capability to a rate store when getting the current
   # cache.
   module CacheableStore
+    include AfterCommitEverywhere
+
     def get_rate(from, to, date = nil)
       return super if date
 
@@ -15,7 +17,9 @@ module ActiveCurrency
     def add_rate(from, to, rate, date = nil)
       super
 
-      Rails.cache.delete(cache_key(from, to))
+      after_commit do
+        Rails.cache.delete(cache_key(from, to))
+      end
     end
 
     private