+++ lib/active_currency/configuration.rb
@@ -5,9 +5,11 @@ module ActiveCurrency
     def initialize
       @remote_bank = :eu_central_bank
       @open_exchange_rates_app_id = nil
+      @multiplier = 1
     end
 
     attr_accessor :remote_bank,
-                  :open_exchange_rates_app_id
+                  :open_exchange_rates_app_id,
+                  :multiplier
   end
 end